Create components & containers directories; move App component to components dir
