+++ src/components/App/index.js
@@ -1,8 +1,8 @@
-import './App.css';
+import styles from './App.module.css';
 
 function App() {
   return (
-    <div className="App">
+    <div className={styles.App}>
       <h1>Property Finder</h1>
     </div>
   );
